Moved files
updated file structure to be component based
